% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{ontology}
\alias{ontology}
\title{Ontological Factors in R}
\usage{
ontology(x = factor(ordered = TRUE), ont_levels, ...)
}
\arguments{
\item{x}{a vector of data.}

\item{ont_levels}{an optional directed \link{igraph} object defining the ontology.}

\item{...}{other parameters passed on to \link{factor}.}
}
\description{
The function \code{ontology} is uset to incode a vector as an ontology,
or a tree-like ordered factor. An \code{ontology} object \emph{is} a factor,
in that it inherits the \code{factor} and \code{ordered} classes, and can therefor
be manipulated as such. The \code{ontology} class adds an attribute, "ont_levels",
that is a directed graph of class \code{igraph}.
}
