% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tct_collapse.R
\name{tct_collapse}
\alias{tct_collapse}
\title{Collapse to a tree level}
\usage{
tct_collapse(x, level, update_levels = TRUE)
}
\arguments{
\item{x}{a \code{tct_tree}.}

\item{level}{a level to collapse.}

\item{update_levels}{whether the \code{levels} and \code{tct_levels} should
be updated to remove descendent levels. Defaults to TRUE.}
}
\description{
Collapse to a tree level, recoding descendent values in the tree to
the collapsed level.
}
\examples{
c("Kyle", "John", "Molly") \%>\%
  tct_tree() \%>\%
  tct_collapse("Kyle")
}
