% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tct_drop.R
\name{tct_drop}
\alias{tct_drop}
\title{Drop unused levels}
\usage{
tct_drop(x, only)
}
\arguments{
\item{x}{a \code{tct_tree}.}

\item{only}{A character vector restricting the set of levels to be dropped.
If supplied, only levels that have no entries and appear in this vector will be removed.}
}
\description{
Compared to base::droplevels(), does not drop NA levels that have values.
}
\examples{
c("Kyle", "John", "Molly") \%>\%
  tct_tree() \%>\%
  tct_drop("Kyle")
}
