% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendents.R
\name{tct_get_descendents}
\alias{tct_get_descendents}
\alias{tct_is_descendent}
\title{Descendents of a tree level}
\usage{
tct_get_descendents(x, level, return_self = TRUE)

tct_is_descendent(x, level, return_self = TRUE)
}
\arguments{
\item{x}{a \code{tct_tree}.}

\item{level}{a level whose descendents will be returned.}

\item{return_self}{whether to return the level in the vector of descendents.}
}
\description{
Get the descendents of a tree level.
}
\section{Functions}{
\itemize{
\item \code{tct_is_descendent}: Returns a logical vector of descendent values.
}}

\examples{
c("Kyle", "John", "Molly") \%>\%
  tct_tree() \%>\%
  tct_get_descendents("Kyle")
}
