% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tct_tree.R
\name{tct_tree}
\alias{tct_tree}
\alias{tct_is_tree}
\alias{is.tct_tree}
\alias{tct_as_tree}
\alias{as.tct_tree}
\alias{tct_levels}
\alias{tct_levels<-}
\title{Tree-like Factors}
\usage{
tct_tree(x = factor(ordered = TRUE), tct_levels, ...)

tct_is_tree(x)

is.tct_tree(x)

tct_as_tree(x)

as.tct_tree(x)

tct_levels(x)

tct_levels(x) <- value
}
\arguments{
\item{x}{a vector of data.}

\item{tct_levels}{an optional directed \link{igraph} object defining the ontology.}

\item{...}{other parameters passed on to \link{factor}.}

\item{value}{a vector of character strings with length at least the number of levels of x.}
}
\description{
The function \code{tct_tree} is used to encode a vector as a tree-like ordered factor.
An \code{tct_tree} object \emph{is} a factor,
in that it inherits the \code{factor} and \code{ordered} classes, and can therefor
be manipulated as such. The \code{tct_tree} class adds an attribute, \code{tct_levels},
that is a directed graph of class \code{igraph}.
}
\details{
\code{tct_is_tree}, \code{is.tct_tree}, \code{tct_as_tree}, and \code{as.tct_tree},
are the membership and coercion functions for the \code{tct_tree} class.

\code{tct_levels} provides access to the \code{tct_levels} attribute of a \code{tct_tree}.
The first form returns the value of the levels of its argument and the second sets the attribute.
}
\section{Functions}{
\itemize{
\item \code{tct_is_tree}: Membership function for the \code{tct_tree} class.

\item \code{is.tct_tree}: Membership function for the \code{tct_tree} class.

\item \code{tct_as_tree}: Coercion function for the \code{tct_tree} class.

\item \code{as.tct_tree}: Coercion function for the \code{tct_tree} class.

\item \code{tct_levels}: Return the value of the \code{tct_levels} attribute of a \code{tct_tree}.

\item \code{tct_levels<-}: Set the value of the \code{tct_levels} attribute of a \code{tct_tree}.
}}

\examples{
c("Kyle", "John", "Molly") \%>\%
  tct_tree()
}
